requirements.txt
# Food Freshness AI - Requirements

# Core
flask>=2.0.0
streamlit>=1.28.0
pillow>=9.0.0

# YOLO Detection
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Gemini AI
google-generativeai>=0.3.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0

# Database
# SQLite is built into Python

# Web scraping for dataset
duckduckgo_search>=3.0.0
requests>=2.28.0

# Excel handling (for legacy storage)
openpyxl>=3.1.0

# Utilities
werkzeug>=2.0.0
python-dotenv>=1.0.0
